requirements.txt
arrow==0.12.1
binaryornot==0.4.4
certifi==2018.8.24
chardet==3.0.4
click==6.7
cookiecutter==1.6.0
django-ipware==2.1.0
future==0.16.0
idna==2.7
Jinja2==2.10
jinja2-time==0.2.0
MarkupSafe==1.0
poyo==0.4.1
python-dateutil==2.7.3
requests==2.19.1
six==1.11.0
urllib3==1.23
whichcraft==0.4.1
